feat(wl-28): Remove Steps component and move Step to UI components
